Create base.txt
Adding the first version of the working script. Note: there are two functions, the first handle all the animation at 100 millisecond intervals. The second function keeps accurate time and writes out the salary value to the pages browser title (usually located in the tab) all when the main page is defocused. The purpose of the second script is to handle the restriction browsers place on javascript loops running faster than one frame a second... we have to keep accurate time somehow AND be able to flying pennies faster than that